fix: Updating playground code
