+++ src/playground/App.tsx
@@ -1,12 +1,12 @@
 import ThemeToggle from './ThemeToggle';
 
-import { HelloWorld } from '@/playground/HelloWorld/HelloWorld';
+import { Playground } from '@/playground';
 
 export const App = () => {
   return (
     <div className='App flex flex-col gap-5 p-4'>
       <ThemeToggle />
-      <HelloWorld />
+      <Playground />
     </div>
   );
 };
